rename `MenuSeminars`->`TopGrid`
